Round two of fixing #124
This time, the path was being canonicalized in another location. My
original verification apparently did not account for this.

All calls to `canonicalize` have been audited, and every location where
such is called, it is also expected that the path exist, and is indeed
considered an error if it does not exist.